+++ CopilotKit/packages/react-core/src/utils/extract.ts
@@ -12,6 +12,7 @@ import {
   TextMessage,
   convertGqlOutputToMessages,
   CopilotRequestType,
+  ForwardedParametersInput,
 } from "@copilotkit/runtime-client-gql";
 import { CopilotContextParams, CopilotMessagesContextParams } from "../context";
 import { defaultCopilotContextCategories } from "../components";
@@ -50,6 +51,7 @@ interface ExtractOptions<T extends Parameter[]> {
   abortSignal?: AbortSignal;
   stream?: (args: StreamHandlerArgs<T>) => void;
   requestType?: CopilotRequestType;
+  forwardedParameters?: ForwardedParametersInput;
 }
 
 interface IncludeOptions {
@@ -66,6 +68,7 @@ export async function extract<const T extends Parameter[]>({
   abortSignal,
   stream,
   requestType = CopilotRequestType.Task,
+  forwardedParameters,
 }: ExtractOptions<T>): Promise<MappedParameterTypes<T>> {
   const { messages } = context;
 
@@ -122,6 +125,7 @@ export async function extract<const T extends Parameter[]>({
           requestType: requestType,
         },
         forwardedParameters: {
+          ...(forwardedParameters ?? {}),
           toolChoice: "function",
           toolChoiceFunctionName: action.name,
         },