+++ CopilotKit/packages/runtime/src/service-adapters/groq/groq-adapter.ts
@@ -102,6 +102,7 @@ export class GroqAdapter implements CopilotServiceAdapter {
       ...(forwardedParameters?.stop && { stop: forwardedParameters.stop }),
       ...(toolChoice && { tool_choice: toolChoice }),
       ...(this.disableParallelToolCalls && { parallel_tool_calls: false }),
+      ...(forwardedParameters?.temperature && { temperature: forwardedParameters.temperature }),
     });
 
     eventSource.stream(async (eventStream$) => {