+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkit/shared
 
+## 1.5.1-next.1
+
+### Patch Changes
+
+- 03f3d6f: Test next prerelease
+
+## 1.5.1-next.0
+
+### Patch Changes
+
+- 27e42d7: testing a prerelease
+
 ## 1.5.0
 
 ### Minor Changes